[REFACTOR] MetadataUtilTest 분리
MetadataUtil에 대한 부분이 MediaRepositoryTest 내부에 있어 분리
